fix(dflash2): unblock DFlash2, and keep this fork's causal small-T attention
DFlash2 could not start at all on this fork. `--spec dflash2` died in CUDA
graph preparation with "context_kv_materialize: invalid cache V".

src/ops/context_kv_materialize/ is a subsystem that arrived whole from
upstream, so nothing conflicted and nothing got reviewed. It required
DType::FP16 for the cyclic cache's V plane while requiring BF16 for K, and
typed cache_v as __half*, storing through
__float2half_rn(__bfloat162float(__float2bfloat16_rn(x))). The consumer of that
cache -- sliding-window attention -- requires BF16 for both planes and takes a
__nv_bfloat16* context_v, so the two could never have worked together: either
the validation rejects it, as it did, or FP16 bits get decoded as BF16. Both
planes are BF16 now, and the store drops the lossy FP16 re-encode (the value was
already rounded through BF16 first, so nothing changes numerically).

This is the same asymmetry that has come back on every catch-up merge, so
materialize.cu now also asserts its plane types against the declared storage via
the new plane_types.h aliases. A new upstream subsystem is the *worst* case for
this class of bug, not the best: no conflict means no review.

Causal small-T attention is reverted to this fork's pre-merge version (12 files
in dense/causal_cache/). Upstream's rework of that subsystem produces
garbage-magnitude output for the whole int8 family on sm_86 -- both geometries,
T=1..16, int8-g64 and rk8v4 alike, roughly 3-8x the reference with a varying
ratio, which is what a softmax denominator taken over the wrong split count
looks like. int8 is this fork's default KV dtype, so correctness there outranks
upstream parity in a path upstream tunes for sm_120. Pre-merge master passes
every one of those cases on this hardware; that is the version restored here.

Also in this commit:
- rk8v4 is routed through the CachePlan overload in the causal-cache fixture.
  The KvCacheStorage overload of make_cache has no packed-int4 branch, so it
  silently built an unpacked INT8 value plane that the kernel then read as
  packed and ran off the end of (access violation). The fixture already
  documented that only the CachePlan overload builds rk8v4; I had missed it.
- The sliding-window test floors its workspace arena at one byte. Direct-route
  shapes legitimately need no split partials, and DeviceArena rejects a zero
  capacity, so run_case(1, 0) aborted the binary with no output at all. Its
  main() now reports exceptions instead of letting them reach terminate.
- The LinearSwiGLU fixture no longer assumes its graph cases ran. On sm_86 the
  A4/A8 graph cases are all refused by the engine and skipped, so the device
  buffer still held the original activation while the final check expected the
  negated one -- reported, misleadingly, as "read-only input was modified".
- apps/cli advertises --spec dflash2 in its usage string; the parser already
  accepted it.

+++ src/ops/context_kv_materialize/context_kv_materialize.cpp
@@ -95,7 +95,9 @@ void validate_cache(const CyclicKVCacheLayerView& cache, std::int32_t padded,
         throw std::invalid_argument("context_kv_materialize: invalid cyclic cache geometry");
     }
     require_tensor(cache.k, DType::BF16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache K");
-    require_tensor(cache.v, DType::FP16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache V");
+    // Symmetric with K. Upstream requires FP16 here; this fork stores both planes as BF16,
+    // and requiring FP16 is what made DFlash2 fail with "invalid cache V".
+    require_tensor(cache.v, DType::BF16, kHeadDim, padded, kKVHeads, lane_capacity, 16, "cache V");
 }
 
 } // namespace
